Update CMPUIKitUtils project location, add property flag to manage the library dependency configuration
