+++ compose/ui/ui-uikit/src/uikitMain/objc/CMPUIKitUtils/CMPUIKitUtils/CMPUIKitUtils.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Android Open Source Project
+ * Copyright 2022 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,12 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.accessibility
+#import <Foundation/Foundation.h>
 
-import androidx.compose.ui.uikit.accessibility.*
-import platform.UIKit.UIView
+//! Project version number for CMPUIKitUtils.
+FOUNDATION_EXPORT double CMPUIKitUtilsVersionNumber;
 
-class AccessibilityElement(
-    private val container: UIView
-): Accessibility() {
-    private val element = CMPAccessibilityElement(container)
+//! Project version string for CMPUIKitUtils.
+FOUNDATION_EXPORT const unsigned char CMPUIKitUtilsVersionString[];
 
-    override fun getPlatform(): String {
-        return "iOS"
-    }
-}
+#import "CMPAccessibilityElement.h"
