+++ gradle.properties
@@ -92,6 +92,9 @@ androidx.validateProjectStructure=false
 jetbrains.compose.jsCompilerTestsEnabled=true
 kotlinWasmEnabled=false
 
+# TODO: Set false when publishing
+useObjcUtilsAsSources=true
+
 # Compose Multiplatform doesn't have its own implementation for Android target.
 # When user adds a dependency on `org.jetbrains.compose.*` in Android target, in fact `android.compose` is used.
 # To know which version, should be used, see compose/frameworks/support/libraryversions.toml
