+++ settings.gradle
@@ -115,11 +115,6 @@ includeProject(":benchmark:benchmark-junit4")
 includeProject(":benchmark:benchmark-macro")
 includeProject(":benchmark:benchmark-macro-junit4")
 includeProject(":benchmark:integration-tests:macrobenchmark-target")
-
-// TODO: Include only when working on a Accessibility Library
-includeProject(":compose:accessibility")
-includeProject(":compose:accessibility:accessibility")
-
 includeProject(":compose:animation")
 includeProject(":compose:animation:animation")
 includeProject(":compose:animation:animation-lint")
@@ -251,6 +246,9 @@ includeProject(":compose:ui:ui-viewbinding")
 includeProject(":compose:ui:ui-viewbinding:ui-viewbinding-samples", "compose/ui/ui-viewbinding/samples")
 includeProject(":compose:ui:ui:integration-tests:ui-demos")
 includeProject(":compose:ui:ui:ui-samples", "compose/ui/ui/samples")
+if (useObjcUtilsAsSources.asBoolean()) {
+    includeProject(":compose:ui:ui-uikit")
+}
 
 includeProject(":lint-checks")
 includeProject(":lint-checks:integration-tests")
